Split IgnoreUserWizard into own file
